dockerize app with postgres

+++ backend/app/schema.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Boolean, Integer, ForeignKey
+from sqlalchemy import Column, String, Boolean, Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 from app.database import Base
 import uuid
@@ -19,8 +19,8 @@ class TodoModel(Base):
     id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
     text = Column(String, nullable=False)
     completed = Column(Boolean, default=False, nullable=False)
-    created_at = Column(Integer, nullable=False)  # Timestamp in milliseconds
-    due_date = Column(Integer, nullable=True)  # Timestamp in milliseconds
+    created_at = Column(BigInteger, nullable=False)  # Timestamp in milliseconds
+    due_date = Column(BigInteger, nullable=True)  # Timestamp in milliseconds
     priority = Column(String, nullable=True)  # "low", "medium", "high"
     category = Column(String, nullable=True)
     user_id = Column(String, ForeignKey("users.id"), nullable=True)
